requirements.txt
et-xmlfile==1.1.0
fpdf==1.7.2
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
six==1.16.0
tzdata==2024.1
